refactor: 세션 리스트 페이지를 react-query를 사용하도록 리팩토링 및 폴더 구조 변경
- pages/SessionListPage 폴더로 분리
